+++ ChangeLog.md
@@ -31,3 +31,12 @@
 - Added ffip like grep commands `util/ffip-search-paths`, `util/ffip-search`,
   `util/ffip-grep-default`, `util/ffip-grep-git-files` and associated functions.
 - Version bump to `0.3.0`.
+
+## [2021-09-09 Thu 01:46]
+- Fixed `util/file-link-re`
+- Changed `util/org-execute-org-heading-max-match-search` to
+  `util/org-execute-customid-or-max-heading-match-search`.
+- Fixed `util/org-get-tree-prop` for not getting `doc-root` when at `doc-root`
+- Added `util/org-insert-citation-to-heading-from-cache`, `util/org-collect-duplicate-customids`,
+  `util/org-helm-show-duplicate-customids`, `util/org-helm-headings` and subroutine for it
+- Version bump to `0.3.1`